fix(libvmaf): eliminate -fsanitize=integer narrowing/overflow in picture.c, libvmaf.c, tensor_io.c (round-5)
Three integer-type defects found by building with `CC=clang -Db_sanitize=integer`:

1. `picture_compute_geometry` (picture.c:125-126): `const int aligned_y/aligned_c`
   computed via `~(DATA_ALIGN - 1)` which evaluates to `int(-64)`, then stored
   into `unsigned` stride fields — implicit signed→unsigned conversion UB. Fix:
   `const unsigned` + `DATA_ALIGN - 1u` mask literal (mirrors picture_cuda.c fix
   from PR #704).

2. `vmaf_init` (libvmaf.c:188): `~cfg.cpumask` (type `uint64_t`) passed to
   `vmaf_set_cpu_flags_mask(unsigned)` with no explicit truncation cast —
   implicit 64→32-bit narrowing. Fix: `(unsigned)(~cfg.cpumask)` with comment
   explaining the deliberate truncation (all CPU flag bits fit in 6 bits).

3. `f16_to_f32_one` subnormal path (tensor_io.c:54,58): normalisation loop used
   `uint32_t exp = 0` and decremented it, wrapping through UINT32_MAX twice to
   arrive at the correct f32 biased exponent by modular arithmetic. Numerically
   correct (verified by test_f16_to_f32_subnormal asserting 2^-24), but the
   unsigned wraps trip -fsanitize=integer. Fix: local `int32_t exp_adj = 1`
   counter bounded to [-9, 1] for any valid 10-bit f16 mantissa.

Fires in: test_picture_alloc_ref_and_unref, test_feature_extractor_flush,
test_speed_qa_flat_input_is_finite, test_anti_dithering_filter,
test_read_pictures_monotonic_accepts_increasing, test_f16_to_f32_subnormal.

All three tests pass clean with zero `runtime error` lines after the fix.
No change to computed scores; tensor_io test asserts bit-exact output.

+++ libvmaf/src/libvmaf.c
@@ -185,7 +185,10 @@ int vmaf_init(VmafContext **vmaf, VmafConfiguration cfg)
     v->cfg = cfg;
 
     vmaf_init_cpu();
-    vmaf_set_cpu_flags_mask(~cfg.cpumask);
+    /* cpumask is uint64_t in the public API; the internal mask is unsigned
+     * (all defined flag bits fit in 32 bits). Truncate explicitly so that
+     * -fsanitize=integer does not fire on the implicit narrowing. */
+    vmaf_set_cpu_flags_mask((unsigned)(~cfg.cpumask));
 
     vmaf_set_log_level(cfg.log_level);
